separate files: config, discord_log, error
